+++ architecture-model-generator/modules/architecture-model-generator-ls-extension/src/main/resources/META-INF/services/org.ballerinalang.langserver.commons.registration.BallerinaClientCapabilitySetter
@@ -1,3 +1,3 @@
 io.ballerina.architecturemodelgenerator.extension.architecture.ArchitectureModelGeneratorClientCapabilitySetter
 io.ballerina.architecturemodelgenerator.extension.persist.PersistERModelGeneratorClientCapabilitySetter
-io.ballerina.desginmodelgenerator.extension.DesignModelGeneratorClientCapabilitiesSetter
+io.ballerina.designmodelgenerator.extension.DesignModelGeneratorClientCapabilitiesSetter
