+++ architecture-model-generator/modules/architecture-model-generator-ls-extension/src/main/resources/META-INF/services/org.ballerinalang.langserver.commons.registration.BallerinaServerCapabilitySetter
@@ -1,3 +1,3 @@
 io.ballerina.architecturemodelgenerator.extension.architecture.ArchitectureModelGeneratorServerCapabilitySetter
 io.ballerina.architecturemodelgenerator.extension.persist.PersistERModelGeneratorServerCapabilitySetter
-io.ballerina.desginmodelgenerator.extension.DesignModelGeneratorServerCapabilitiesSetter
+io.ballerina.designmodelgenerator.extension.DesignModelGeneratorServerCapabilitiesSetter
