+++ architecture-model-generator/modules/architecture-model-generator-ls-extension/src/main/resources/META-INF/services/org.ballerinalang.langserver.commons.service.spi.ExtendedLanguageServerService
@@ -1,3 +1,3 @@
 io.ballerina.architecturemodelgenerator.extension.architecture.ArchitectureModelGeneratorService
 io.ballerina.architecturemodelgenerator.extension.persist.PersistERModelGeneratorService
-io.ballerina.desginmodelgenerator.extension.DesignModelGeneratorService
+io.ballerina.designmodelgenerator.extension.DesignModelGeneratorService
